requirements.txt
keras
numpy
Pillow
scikit-image
scikit-learn
tensorflow

